Use proper serialization and fix [object Object] when working with matrices
